revset: move `ReverseRevsetGraphIterator` into `revset` module
The iterator is not specific to the implementation in
`revset_graph_iterator`, so it belongs in the standard `revset`
module.
