+++ lib/src/revset.rs
@@ -1488,6 +1488,49 @@ pub struct RevsetWorkspaceContext<'a> {
     pub workspace_root: &'a Path,
 }
 
+pub struct ReverseRevsetGraphIterator<'index> {
+    items: Vec<(IndexEntry<'index>, Vec<RevsetGraphEdge>)>,
+}
+
+impl<'index> ReverseRevsetGraphIterator<'index> {
+    pub fn new<'revset>(
+        input: Box<dyn Iterator<Item = (IndexEntry<'index>, Vec<RevsetGraphEdge>)> + 'revset>,
+    ) -> Self {
+        let mut entries = vec![];
+        let mut reverse_edges: HashMap<IndexPosition, Vec<RevsetGraphEdge>> = HashMap::new();
+        for (entry, edges) in input {
+            for RevsetGraphEdge { target, edge_type } in edges {
+                reverse_edges
+                    .entry(target)
+                    .or_default()
+                    .push(RevsetGraphEdge {
+                        target: entry.position(),
+                        edge_type,
+                    })
+            }
+            entries.push(entry);
+        }
+
+        let mut items = vec![];
+        for entry in entries.into_iter() {
+            let edges = reverse_edges
+                .get(&entry.position())
+                .cloned()
+                .unwrap_or_default();
+            items.push((entry, edges));
+        }
+        Self { items }
+    }
+}
+
+impl<'index> Iterator for ReverseRevsetGraphIterator<'index> {
+    type Item = (IndexEntry<'index>, Vec<RevsetGraphEdge>);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        self.items.pop()
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
